another VMBridge cleanup step (JDK 1.8 is the minimum at the moment)
