+++ src/org/mozilla/javascript/JavaAdapter.java
@@ -397,8 +397,7 @@ public final class JavaAdapter implements IdFunctionCall
             for (int j = 0; j < methods.length; j++) {
                 Method method = methods[j];
                 int mods = method.getModifiers();
-                if (Modifier.isStatic(mods) || Modifier.isFinal(mods) ||
-                    VMBridge.instance.isDefaultMethod(method)) {
+                if (Modifier.isStatic(mods) || Modifier.isFinal(mods) || method.isDefault()) {
                     continue;
                 }
                 String methodName = method.getName();
